requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-dotenv
jinja2
python-multipart
